+++ Source/TresGame/Public/TresPhysObjSkeletalMeshActor.h
@@ -15,22 +15,22 @@ class ATresPhysObjSkeletalMeshActor : public ASkeletalMeshActor {
 public:
 protected:
     UPROPERTY(BlueprintReadWrite, EditAnywhere)
-    uint8 m_bDisappearAfterWake: 1;
+    uint32 m_bDisappearAfterWake: 1;
     
     UPROPERTY(BlueprintReadWrite, EditAnywhere)
-    uint8 m_bDisappearLeaveInitLoc: 1;
+    uint32 m_bDisappearLeaveInitLoc: 1;
     
     UPROPERTY(BlueprintReadWrite, EditAnywhere)
-    uint8 m_bDisappearLeaveInitPhysVolume: 1;
+    uint32 m_bDisappearLeaveInitPhysVolume: 1;
     
     UPROPERTY(BlueprintReadWrite, EditAnywhere)
-    uint8 m_bDisappearInnerWaterPhysVolume: 1;
+    uint32 m_bDisappearInnerWaterPhysVolume: 1;
     
     UPROPERTY(BlueprintReadWrite, EditAnywhere)
-    uint8 m_bDisappearInnerOceanPhysVolume: 1;
+    uint32 m_bDisappearInnerOceanPhysVolume: 1;
     
     UPROPERTY(BlueprintReadWrite, EditAnywhere)
-    uint8 m_bDisappearOnTakeDamage: 1;
+    uint32 m_bDisappearOnTakeDamage: 1;
     
     UPROPERTY(BlueprintReadWrite, EditAnywhere)
     float m_DisappearAfterWakeTimeSetting;
@@ -42,7 +42,7 @@ protected:
     int32 m_DisappearOnTakeDamageSetting;
     
     UPROPERTY(BlueprintReadWrite, EditAnywhere)
-    uint8 m_bAutoDisappearEffectProc: 1;
+    uint32 m_bAutoDisappearEffectProc: 1;
     
     UPROPERTY(BlueprintReadWrite, EditAnywhere)
     float m_DisappearFadeSetting;
