+++ Source/TresGame/Public/TresRestrictActionVolume.h
@@ -9,58 +9,58 @@ class TRESGAME_API ATresRestrictActionVolume : public ATresVolume {
 public:
 protected:
     UPROPERTY(EditAnywhere)
-    uint8 m_TownArea: 1;
+    uint32 m_TownArea: 1;
     
     UPROPERTY(EditAnywhere)
-    uint8 m_RestrictATTACK: 1;
+    uint32 m_RestrictATTACK: 1;
     
     UPROPERTY(EditAnywhere)
-    uint8 m_RestrictMAGIC: 1;
+    uint32 m_RestrictMAGIC: 1;
     
     UPROPERTY(EditAnywhere)
-    uint8 m_RestrictITEM: 1;
+    uint32 m_RestrictITEM: 1;
     
     UPROPERTY(EditAnywhere)
-    uint8 m_RestrictSITUATION: 1;
+    uint32 m_RestrictSITUATION: 1;
     
     UPROPERTY(EditAnywhere)
-    uint8 m_RestrictLINK: 1;
+    uint32 m_RestrictLINK: 1;
     
     UPROPERTY(EditAnywhere)
-    uint8 m_RestrictATTRACTIONFLOW: 1;
+    uint32 m_RestrictATTRACTIONFLOW: 1;
     
     UPROPERTY(EditAnywhere)
-    uint8 m_RestrictSTYLE: 1;
+    uint32 m_RestrictSTYLE: 1;
     
     UPROPERTY(EditAnywhere)
-    uint8 m_RestrictFRIENDLINK: 1;
+    uint32 m_RestrictFRIENDLINK: 1;
     
     UPROPERTY(EditAnywhere)
-    uint8 m_RestrictGUARD: 1;
+    uint32 m_RestrictGUARD: 1;
     
     UPROPERTY(EditAnywhere)
-    uint8 m_RestrictJUMP: 1;
+    uint32 m_RestrictJUMP: 1;
     
     UPROPERTY(EditAnywhere)
-    uint8 m_RestrictSHOOTLOCK: 1;
+    uint32 m_RestrictSHOOTLOCK: 1;
     
     UPROPERTY(EditAnywhere)
-    uint8 m_RestrictRUN: 1;
+    uint32 m_RestrictRUN: 1;
     
     UPROPERTY(EditAnywhere)
-    uint8 m_RestrictDIVEFALL: 1;
+    uint32 m_RestrictDIVEFALL: 1;
     
     UPROPERTY(EditAnywhere)
-    uint8 m_RestrictDeepDiving: 1;
+    uint32 m_RestrictDeepDiving: 1;
     
     UPROPERTY(EditAnywhere)
-    uint8 m_RestrictCancelSituation: 1;
+    uint32 m_RestrictCancelSituation: 1;
     
     UPROPERTY(EditAnywhere)
-    uint8 m_RestrictAlwaysUseSpawnPoint: 1;
+    uint32 m_RestrictAlwaysUseSpawnPoint: 1;
     
     UPROPERTY(EditAnywhere)
-    uint8 m_RestrictPhotoMode: 1;
+    uint32 m_RestrictPhotoMode: 1;
     
 public:
     ATresRestrictActionVolume();
